# frontend/specs/openapi.json
{"openapi":"3.0.0","paths":{"/api/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":"User successfully registered","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponseDto"}}}},"409":{"description":"User with this email or username already exists"}},"summary":"Register a new user","tags":["auth"]}},"/api/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"User successfully logged in","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponseDto"}}}},"401":{"description":"Invalid credentials"}},"summary":"Login with email and password","tags":["auth"]}},"/api/auth/reset-password":{"post":{"operationId":"AuthController_resetPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordDto"}}}},"responses":{"200":{"description":"Password reset instructions sent"},"404":{"description":"User not found"}},"summary":"Request password reset","tags":["auth"]}},"/api/auth/change-password":{"post":{"operationId":"AuthController_changePassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordDto"}}}},"responses":{"200":{"description":"Password changed successfully"},"401":{"description":"Invalid or expired token"}},"summary":"Change password with token","tags":["auth"]}},"/api/users":{"get":{"operationId":"UserController_findAll","parameters":[],"responses":{"200":{"description":"Return all users","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponseDto"}}}}}},"summary":"Get all users","tags":["users"]}},"/api/users/leaderboard":{"get":{"operationId":"UserController_getLeaderboard","parameters":[{"name":"page","required":false,"in":"query","description":"Page number","schema":{"type":"number"}},{"name":"limit","required":false,"in":"query","description":"Number of items per page","schema":{"type":"number"}}],"responses":{"200":{"description":"Return the leaderboard","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LeaderboardDto"}}}}},"summary":"Get user leaderboard","tags":["users"]}},"/api/users/me":{"get":{"operationId":"UserController_getProfile","parameters":[],"responses":{"200":{"description":"Return the current user","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}},"401":{"description":"Unauthorized"}},"security":[{"bearer":[]}],"summary":"Get current user profile","tags":["users"]},"put":{"operationId":"UserController_updateProfile","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"Return the updated user","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}},"401":{"description":"Unauthorized"},"404":{"description":"User not found"}},"security":[{"bearer":[]}],"summary":"Update current user profile","tags":["users"]}},"/api/users/username/{username}":{"get":{"operationId":"UserController_findByUsername","parameters":[{"name":"username","required":true,"in":"path","description":"The username of the user","schema":{"type":"string"}}],"responses":{"200":{"description":"Return the user","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}},"404":{"description":"User not found"}},"summary":"Get user by username","tags":["users"]}},"/api/memecoins":{"get":{"operationId":"MemecoinController_findAll","parameters":[{"name":"page","required":false,"in":"query","description":"Page number","schema":{"type":"number"}},{"name":"limit","required":false,"in":"query","description":"Number of items per page","schema":{"type":"number"}},{"name":"sortBy","required":false,"in":"query","description":"Sort by field","schema":{"enum":["createdAt","name","symbol","totalSupply"],"type":"string"}},{"name":"order","required":false,"in":"query","description":"Sort order","schema":{"enum":["ASC","DESC"],"type":"string"}}],"responses":{"200":{"description":"Return all memecoins","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MemecoinResponseDto"}}}}}},"summary":"Get all memecoins","tags":["memecoins"]},"post":{"operationId":"MemecoinController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMemecoinDto"}}}},"responses":{"201":{"description":"Memecoin successfully created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MemecoinResponseDto"}}}},"400":{"description":"Invalid input or insufficient balance"},"401":{"description":"Unauthorized"},"409":{"description":"Memecoin with this name or symbol already exists"}},"security":[{"bearer":[]}],"summary":"Create a new memecoin","tags":["memecoins"]}},"/api/memecoins/{id}":{"get":{"operationId":"MemecoinController_findOne","parameters":[{"name":"id","required":true,"in":"path","description":"The ID of the memecoin","schema":{"type":"string"}}],"responses":{"200":{"description":"Return the memecoin","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MemecoinResponseDto"}}}},"404":{"description":"Memecoin not found"}},"summary":"Get memecoin by ID","tags":["memecoins"]}},"/api/memecoins/symbol/{symbol}":{"get":{"operationId":"MemecoinController_findBySymbol","parameters":[{"name":"symbol","required":true,"in":"path","description":"The symbol of the memecoin","schema":{"type":"string"}}],"responses":{"200":{"description":"Return the memecoin","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MemecoinResponseDto"}}}},"404":{"description":"Memecoin not found"}},"summary":"Get memecoin by symbol","tags":["memecoins"]}},"/api/memecoins/{id}/transactions":{"get":{"operationId":"MemecoinController_getTransactions","parameters":[{"name":"id","required":true,"in":"path","description":"The ID of the memecoin","schema":{"type":"string"}}],"responses":{"200":{"description":"Return the memecoin transactions"},"404":{"description":"Memecoin not found"}},"summary":"Get memecoin transactions","tags":["memecoins"]}},"/api/statistics/trading-volume":{"get":{"operationId":"StatisticsController_getTradingVolume","parameters":[{"name":"timeframe","required":false,"in":"query","description":"Timeframe for statistics","schema":{"enum":["24h","7d","30d"],"type":"string"}},{"name":"memecoinId","required":false,"in":"query","description":"Filter by memecoin ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Return trading volume statistics","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TradingVolumeDto"}}}}},"summary":"Get trading volume statistics","tags":["statistics"]}},"/api/wallet":{"get":{"operationId":"WalletController_getWallet","parameters":[],"responses":{"200":{"description":"Return the user's wallet","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WalletResponseDto"}}}},"401":{"description":"Unauthorized"},"404":{"description":"Wallet not found"}},"security":[{"bearer":[]}],"summary":"Get current user's wallet","tags":["wallet"]}},"/api/wallet/transactions":{"get":{"operationId":"WalletController_getTransactions","parameters":[{"name":"page","required":false,"in":"query","description":"Page number","schema":{"type":"number"}},{"name":"limit","required":false,"in":"query","description":"Number of items per page","schema":{"type":"number"}}],"responses":{"200":{"description":"Return the user's transaction history","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TransactionResponseDto"}}}}},"401":{"description":"Unauthorized"}},"security":[{"bearer":[]}],"summary":"Get current user's transaction history","tags":["wallet"]}},"/api/trading/buy":{"post":{"operationId":"TradingController_buyMemecoin","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BuyMemecoinDto"}}}},"responses":{"201":{"description":"Memecoin successfully purchased","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TradeResponseDto"}}}},"400":{"description":"Invalid input or insufficient balance"},"401":{"description":"Unauthorized"},"404":{"description":"Memecoin or user not found"}},"security":[{"bearer":[]}],"summary":"Buy memecoin","tags":["trading"]}},"/api/trading/sell":{"post":{"operationId":"TradingController_sellMemecoin","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SellMemecoinDto"}}}},"responses":{"201":{"description":"Memecoin successfully sold","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TradeResponseDto"}}}},"400":{"description":"Invalid input or insufficient memecoin balance"},"401":{"description":"Unauthorized"},"404":{"description":"Memecoin or user not found"}},"security":[{"bearer":[]}],"summary":"Sell memecoin","tags":["trading"]}}},"info":{"title":"Zero to Hundred API","description":"The Zero to Hundred memecoin platform API","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"RegisterDto":{"type":"object","properties":{"username":{"type":"string","description":"The username of the user","example":"johndoe"},"email":{"type":"string","description":"The email of the user","example":"user@example.com"},"password":{"type":"string","description":"The password of the user","example":"password123"},"fullName":{"type":"string","description":"The full name of the user","example":"John Doe"}},"required":["username","email","password"]},"AuthResponseDto":{"type":"object","properties":{"accessToken":{"type":"string","description":"The JWT access token","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."},"userId":{"type":"string","description":"The user ID","example":"123e4567-e89b-12d3-a456-426614174000"},"username":{"type":"string","description":"The username","example":"johndoe"},"email":{"type":"string","description":"The user email","example":"user@example.com"}},"required":["accessToken","userId","username","email"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","description":"The email of the user","example":"user@example.com"},"password":{"type":"string","description":"The password of the user","example":"password123"}},"required":["email","password"]},"ResetPasswordDto":{"type":"object","properties":{"email":{"type":"string","description":"The email of the user","example":"user@example.com"}},"required":["email"]},"ChangePasswordDto":{"type":"object","properties":{"token":{"type":"string","description":"The reset token","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."},"password":{"type":"string","description":"The new password","example":"newPassword123"},"confirmPassword":{"type":"string","description":"Confirm the new password","example":"newPassword123"}},"required":["token","password","confirmPassword"]},"UserResponseDto":{"type":"object","properties":{"id":{"type":"string","description":"The unique identifier of the user"},"username":{"type":"string","description":"The username of the user"},"email":{"type":"string","description":"The email of the user"},"fullName":{"type":"string","description":"The full name of the user"},"profilePictureUrl":{"type":"string","description":"The profile picture URL of the user"},"bannerUrl":{"type":"string","description":"The banner URL of the user"},"description":{"type":"string","description":"The description of the user"},"backgroundColor":{"type":"string","description":"The background color of the user profile"},"textColor":{"type":"string","description":"The text color of the user profile"},"isActive":{"type":"boolean","description":"Whether the user is active"},"createdAt":{"format":"date-time","type":"string","description":"The date when the user was created"},"updatedAt":{"format":"date-time","type":"string","description":"The date when the user was last updated"}},"required":["id","username","email","fullName","profilePictureUrl","bannerUrl","description","backgroundColor","textColor","isActive","createdAt","updatedAt"]},"LeaderboardItemDto":{"type":"object","properties":{"user":{"description":"The user","allOf":[{"$ref":"#/components/schemas/UserResponseDto"}]},"rank":{"type":"number","description":"The rank of the user in the leaderboard"},"zthBalance":{"type":"number","description":"The ZTH balance of the user"}},"required":["user","rank","zthBalance"]},"LeaderboardDto":{"type":"object","properties":{"leaderboard":{"description":"The leaderboard items","type":"array","items":{"$ref":"#/components/schemas/LeaderboardItemDto"}},"total":{"type":"number","description":"The total number of users in the leaderboard"}},"required":["leaderboard","total"]},"UpdateUserDto":{"type":"object","properties":{"username":{"type":"string","description":"The username of the user","example":"johndoe"},"fullName":{"type":"string","description":"The full name of the user","example":"John Doe"},"profilePictureUrl":{"type":"string","description":"The profile picture URL of the user","example":"https://example.com/profile.jpg"},"bannerUrl":{"type":"string","description":"The banner URL of the user","example":"https://example.com/banner.jpg"},"description":{"type":"string","description":"The description of the user","example":"I am a memecoin enthusiast"},"backgroundColor":{"type":"string","description":"The background color of the user profile","example":"#f5f5f5"},"textColor":{"type":"string","description":"The text color of the user profile","example":"#333333"}}},"MemecoinResponseDto":{"type":"object","properties":{"id":{"type":"string","description":"The unique identifier of the memecoin"},"name":{"type":"string","description":"The name of the memecoin"},"symbol":{"type":"string","description":"The symbol of the memecoin"},"description":{"type":"string","description":"The description of the memecoin"},"logoUrl":{"type":"string","description":"The logo URL of the memecoin"},"creator":{"description":"The creator of the memecoin","allOf":[{"$ref":"#/components/schemas/UserResponseDto"}]},"creatorId":{"type":"string","description":"The ID of the creator"},"totalSupply":{"type":"string","description":"The total supply of the memecoin"},"currentPrice":{"type":"string","description":"The current price of the memecoin"},"marketCap":{"type":"string","description":"The market cap of the memecoin"},"volume24h":{"type":"string","description":"The 24h trading volume of the memecoin"},"createdAt":{"format":"date-time","type":"string","description":"The date when the memecoin was created"},"updatedAt":{"format":"date-time","type":"string","description":"The date when the memecoin was last updated"}},"required":["id","name","symbol","description","logoUrl","creator","creatorId","totalSupply","currentPrice","marketCap","volume24h","createdAt","updatedAt"]},"CreateMemecoinDto":{"type":"object","properties":{"name":{"type":"string","description":"The name of the memecoin","example":"Doge Coin"},"symbol":{"type":"string","description":"The symbol of the memecoin","example":"DOGE"},"description":{"type":"string","description":"The description of the memecoin","example":"A fun cryptocurrency featuring the Shiba Inu dog from the \"Doge\" meme."},"logoUrl":{"type":"string","description":"The logo URL of the memecoin","example":"https://example.com/doge-logo.png"}},"required":["name","symbol"]},"MemecoinVolumeDto":{"type":"object","properties":{"id":{"type":"string","description":"The ID of the memecoin"},"ticker":{"type":"string","description":"The ticker symbol of the memecoin"},"volume":{"type":"string","description":"The trading volume for this memecoin in ZTH"}},"required":["id","ticker","volume"]},"TradingVolumeDto":{"type":"object","properties":{"totalVolume":{"type":"string","description":"The total trading volume in ZTH"},"buyVolume":{"type":"string","description":"The buy volume in ZTH"},"sellVolume":{"type":"string","description":"The sell volume in ZTH"},"timeframe":{"type":"string","description":"The timeframe for the statistics","enum":["24h","7d","30d"]},"memecoins":{"description":"Volume statistics per memecoin","type":"array","items":{"$ref":"#/components/schemas/MemecoinVolumeDto"}}},"required":["totalVolume","buyVolume","sellVolume","timeframe","memecoins"]},"WalletHoldingResponseDto":{"type":"object","properties":{"id":{"type":"string","description":"The unique identifier of the wallet holding"},"walletId":{"type":"string","description":"The ID of the wallet"},"memecoin":{"description":"The memecoin that is held","allOf":[{"$ref":"#/components/schemas/MemecoinResponseDto"}]},"memecoinId":{"type":"string","description":"The ID of the memecoin"},"amount":{"type":"string","description":"The amount of the memecoin held"},"createdAt":{"format":"date-time","type":"string","description":"The date when the holding was created"},"updatedAt":{"format":"date-time","type":"string","description":"The date when the holding was last updated"}},"required":["id","walletId","memecoin","memecoinId","amount","createdAt","updatedAt"]},"WalletResponseDto":{"type":"object","properties":{"id":{"type":"string","description":"The unique identifier of the wallet"},"zthBalance":{"type":"string","description":"The ZTH balance of the wallet"},"ownerId":{"type":"string","description":"The ID of the owner"},"owner":{"description":"The owner of the wallet","allOf":[{"$ref":"#/components/schemas/UserResponseDto"}]},"holdings":{"description":"The holdings of the wallet","type":"array","items":{"$ref":"#/components/schemas/WalletHoldingResponseDto"}},"createdAt":{"format":"date-time","type":"string","description":"The date when the wallet was created"},"updatedAt":{"format":"date-time","type":"string","description":"The date when the wallet was last updated"}},"required":["id","zthBalance","ownerId","owner","holdings","createdAt","updatedAt"]},"TransactionResponseDto":{"type":"object","properties":{"id":{"type":"string","description":"The unique identifier of the transaction"},"type":{"type":"string","description":"The type of transaction","enum":["BUY","SELL","CREATE"]},"memeCoinAmount":{"type":"string","description":"The amount of the memecoin involved in the transaction"},"zthAmount":{"type":"string","description":"The total value of the transaction in ZTH"},"price":{"type":"string","description":"The price of the memecoin at the time of the transaction"},"user":{"description":"The user who performed the transaction","allOf":[{"$ref":"#/components/schemas/UserResponseDto"}]},"userId":{"type":"string","description":"The ID of the user"},"memecoin":{"description":"The memecoin involved in the transaction","allOf":[{"$ref":"#/components/schemas/MemecoinResponseDto"}]},"memecoinId":{"type":"string","description":"The ID of the memecoin"},"createdAt":{"format":"date-time","type":"string","description":"The date when the transaction was created"}},"required":["id","type","memeCoinAmount","zthAmount","price","user","userId","memecoin","memecoinId","createdAt"]},"BuyMemecoinDto":{"type":"object","properties":{"memecoinId":{"type":"string","description":"The ID of the memecoin to buy","example":"123e4567-e89b-12d3-a456-426614174000"},"amount":{"type":"string","description":"The amount of ZTH to spend","example":"10"},"requestPrice":{"type":"string","description":"The price of the memecoin at the time of the request","example":"0.1"},"slippageTolerance":{"type":"number","description":"The maximum slippage tolerance in percentage","example":"2.5","default":"1"}},"required":["memecoinId","amount","requestPrice"]},"TradeResponseDto":{"type":"object","properties":{"transaction":{"description":"The transaction details","allOf":[{"$ref":"#/components/schemas/TransactionResponseDto"}]},"memecoin":{"description":"The memecoin details","allOf":[{"$ref":"#/components/schemas/MemecoinResponseDto"}]},"walletHolding":{"description":"The wallet holding memecoin","allOf":[{"$ref":"#/components/schemas/WalletHoldingResponseDto"}]}},"required":["transaction","memecoin","walletHolding"]},"SellMemecoinDto":{"type":"object","properties":{"memecoinId":{"type":"string","description":"The ID of the memecoin to sell","example":"123e4567-e89b-12d3-a456-426614174000"},"amount":{"type":"string","description":"The amount of memecoin tokens to sell","example":"5"},"requestPrice":{"type":"string","description":"The price of the memecoin at the time of the request","example":"0.1"},"slippageTolerance":{"type":"number","description":"The maximum slippage tolerance in percentage","example":"2.5","default":"1"}},"required":["memecoinId","amount","requestPrice"]}}}}